+++ src/ga/abzzezz/game/maingame/gui/screens/LevelBuilder.java
@@ -15,8 +15,8 @@ import ga.abzzezz.game.maingame.gui.basis.TextBox;
 import ga.abzzezz.game.maingame.object.Prevent;
 import ga.abzzezz.game.maingame.object.impl.Block;
 import ga.abzzezz.game.maingame.utility.ColorHelper;
-import org.joml.Vector2i;
 import org.lwjgl.input.Keyboard;
+import org.lwjgl.util.vector.Vector2f;
 
 import java.awt.*;
 import java.util.ArrayList;
@@ -49,9 +49,9 @@ public class LevelBuilder extends GuiScreen {
             /*
             Add object to prevents list, so it can be stored etc.
              */
-            prevents.add(new Block("B" + System.currentTimeMillis(), new Vector2i(display()[0] / 2, display()[1] / 2), 50, 50, Color.RED));
+            prevents.add(new Block("B" + System.currentTimeMillis(), new Vector2f(display()[0] / 2, display()[1] / 2), 50, 50, Color.RED));
         } else if (buttonID == 1) {
-            prevents.add(new Block("Player", new Vector2i(display()[0] / 2, display()[1] / 2), 30, 30, Color.GREEN));
+            prevents.add(new Block("Player", new Vector2f(display()[0] / 2, display()[1] / 2), 30, 30, Color.GREEN));
         } else if (buttonID == 3) {
             Main.getMain().getLevelSystem().saveLevel(prevents);
         } else if (buttonID == 4) {
